dev: modularize parser tests
